+++ lib/widgets/home/feed/post_action_bar.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:phosphor_flutter/phosphor_flutter.dart';
-import 'package:stepp_app/constants/feed/post_box_size.dart';
-import 'package:stepp_app/constants/feed/post_box_ui_strings.dart';
+import 'package:stepp_app/constants/home/feed/post_box_size.dart';
+import 'package:stepp_app/constants/home/feed/post_box_ui_strings.dart';
 import 'package:stepp_app/utils/build_context_helper.dart';
 
 class PostActionBar extends StatefulWidget {
