+++ lib/widgets/home/feed/post_box.dart
@@ -2,10 +2,10 @@ import 'package:flutter/material.dart';
 import 'package:stepp_app/constants/mock_data.dart';
 import 'package:stepp_app/constants/routes.dart';
 import 'package:stepp_app/data_models/feed/post.dart';
-import 'package:stepp_app/widgets/feed/post_action_bar.dart';
-import 'package:stepp_app/widgets/feed/post_content.dart';
-import 'package:stepp_app/widgets/feed/post_header_bar.dart';
-import 'package:stepp_app/widgets/feed/post_image.dart';
+import 'package:stepp_app/widgets/home/feed/post_action_bar.dart';
+import 'package:stepp_app/widgets/home/feed/post_content.dart';
+import 'package:stepp_app/widgets/home/feed/post_header_bar.dart';
+import 'package:stepp_app/widgets/home/feed/post_image.dart';
 
 class PostBox extends StatefulWidget {
   final PostModel post;
