+++ lib/widgets/home/feed/post_image.dart
@@ -1,6 +1,6 @@
 import 'package:cached_network_image/cached_network_image.dart';
 import 'package:flutter/material.dart';
-import 'package:stepp_app/constants/feed/post_box_size.dart';
+import 'package:stepp_app/constants/home/feed/post_box_size.dart';
 import 'package:stepp_app/constants/place/stepp_place_ui_strings.dart';
 import 'package:stepp_app/styles/app_theme.dart';
 import 'package:stepp_app/utils/build_context_helper.dart';
